refactor(opal): inline SidebarWrapper + fix onFoldedChange side-effect
- Inline SidebarWrapper directly into sidebar/components.tsx and
  delete the separate file (fewer files, related code co-located)
- Move onFoldedChange notification out of the state updater into a
  useEffect so the updater is pure and safe under StrictMode
- Drop the @/components/resizable/constants import from AppProvider by
  inlining the cookie name string directly
